Refactor Mapview
