+++ backend/apps/mapview/views/__init__.py
@@ -1,0 +1,3 @@
+from match4healthcare.utils.imports import import_submodules
+
+import_submodules(globals(), __name__, __path__)
